outposts/ldap: fix race condition when refreshing the provider
Fixes the race condition causing the crash found in #4138, which doesn't
actually have anything to do with the issue itself.

As far as I can work out, when the outpost refreshes its list of
providers, it copies over its `boundUsers`, probably to avoid having to
fetch them all again, and does so by making a shallow copy of that
`map`, but not the mutex associated with it. It now has multiple
references to the same map, each protected by a different mutex, which
under certain conditions can cause a `concurrent map read and map write`
error.

The fix replaces the duo map + mutex with a reference to a sync.Map
which does basically the same thing, but internally. We now don't need
to worry about properly copying the mutex and such, as it is handled by
sync.Map.

Signed-off-by: Marc 'risson' Schmitt <marc.schmitt@risson.space>